renamed files (no longer 18 dof)
