+++ nmf_gym/__init__.py
@@ -2,11 +2,11 @@ from gym.envs.registration import register
 from importlib_metadata import entry_points
 
 register(
-    id='nmf18-simple_position_control-v0',
-    entry_point='nmf_gym.envs:NMF18SimplePositionControlEnv'
+    id='nmf-simple_position_control-v0',
+    entry_point='nmf_gym.envs:NMFSimplePositionControlEnv'
 )
 
 register(
-    id='nmf18-pos2pos_distance-v0',
-    entry_point='nmf_gym.envs:NMF18Pos2PosDistanceEnv'
+    id='nmf-pos2pos_distance-v0',
+    entry_point='nmf_gym.envs:NMFPos2PosDistanceEnv'
 )
